Split runtime_bridge.py into the webchirp_bridge package
The 2,850-line Python runtime is now a package of thirteen modules, one per
concern: chirp_loader (lazy CHIRP import hook, driver enumeration),
serial_pipe (pyserial stand-in over Web Serial), clone (download/upload
sessions), driver_cache (cached clone image and radios built from it),
channel_rows, power_levels, row_validation, radio_memories, radio_settings,
column_metadata, images, jsbridge and runtime_errors. Every definition moved
verbatim with its comments; the module graph is acyclic and each module
imports what it uses.

runtime_bridge.py stays the entry point that seedPyodideRuntime() executes
into Pyodide's globals. It now imports the package and flattens each
module's namespace into the globals, so the RPC expressions and the test
snippets keep resolving names as before. Tests that swapped a JS callable by
rebinding a global now patch the owning module's attribute instead, since a
module resolves names in its own namespace.

The browser fetches the package file by file. The URL table lives in
runtime-rpc.js because build-dist.mjs rewrites references to hashed names in
.js files only and copies .mjs files verbatim; the provider refuses to
construct if a listed file lacks a URL, and test-build-dist.mjs pins the
pairing statically. chirp.drivers is imported explicitly before the finder is
installed so it stays a namespace package regardless of module import order.

check:python compiles the whole web/python tree and pyright resolves the
package through extraPaths.

Model: Claude Fable 5.1 (claude-fable-5-1)
