+++ package.json
@@ -14,7 +14,7 @@
     "radio:write": "node --experimental-wasm-stack-switching scripts/radio-codeplug.mjs write",
     "test": "npm run check:syntax && npm run check:types && npm run test:channels && npm run test:webusb && npm run test:settings && npm run test:build",
     "check:syntax": "find web scripts -name '*.js' -o -name '*.mjs' | xargs -n1 node --check",
-    "check:python": "python3 -m py_compile web/python/runtime_bridge.py",
+    "check:python": "python3 -m compileall -q web/python",
     "check:types": "pyright",
     "test:channels": "node --test --experimental-wasm-stack-switching scripts/test-channel-list.mjs scripts/test-channel-values.mjs scripts/test-ui-radio-loading.mjs scripts/test-call-queue.mjs scripts/test-runtime-bootstrap.mjs scripts/test-driver-import-race.mjs scripts/test-chirp-import-errors.mjs scripts/test-clipboard.mjs scripts/test-ui-channel-cut.mjs scripts/test-ui-channel-locations.mjs scripts/test-export-filename.mjs scripts/test-image-metadata.mjs scripts/test-datasources.mjs scripts/test-staticmap.mjs scripts/test-ui-repeater-map.mjs scripts/test-ui-repeater-query.mjs scripts/test-ui-repeater-modal.mjs scripts/test-ui-query-fields.mjs scripts/test-rsgb.mjs scripts/test-ui-file-drop.mjs scripts/test-dom-selectors.mjs scripts/test-debug-log.mjs scripts/test-analytics.mjs scripts/test-analytics-events.mjs scripts/test-sentry.mjs scripts/test-ga-dimensions.mjs scripts/test-gettext-builtins.mjs scripts/test-issue-report.mjs scripts/test-csv-export-parity.mjs scripts/test-power-level-watts.mjs scripts/test-channel-round-trip.mjs scripts/test-channel-extras.mjs scripts/test-memory-comments.mjs scripts/test-unreadable-channels.mjs scripts/test-pyserial-shim.mjs scripts/test-serial-in-waiting.mjs scripts/test-metadataless-image-load.mjs scripts/test-progress.mjs scripts/test-clone-progress.mjs scripts/test-clone-button-gating.mjs scripts/test-clone-baud-plumbing.mjs scripts/test-clone-detect-from-serial.mjs",
     "test:webusb": "node --test scripts/test-webusb-serial.mjs scripts/test-ftdi-webusb.mjs scripts/test-pl2303-webusb.mjs scripts/test-ch340-webusb.mjs scripts/test-cp2102-webusb.mjs scripts/test-serial-polyfill.mjs scripts/test-serial-port-cancel.mjs scripts/test-loopback-suite.mjs scripts/test-serial-test-page.mjs scripts/test-loopback-issue.mjs scripts/test-port-loss.mjs scripts/test-clone-baud-rate.mjs scripts/test-serial-control-lines.mjs scripts/test-serial-reconfigure.mjs scripts/test-cli-serial-bridge.mjs",
